add delivery business logic and change Double to BigDecimal in weight, volume, height, depth fields

+++ commerce/interaction-api/src/main/java/ru/yandex/practicum/interaction/dto/order/OrderDto.java
@@ -38,10 +38,10 @@ public class OrderDto {
     OrderState state;
 
     // Общий вес доставки.
-    Double deliveryWeight;
+    BigDecimal deliveryWeight;
 
     // Общий объём доставки.
-    Double deliveryVolume;
+    BigDecimal deliveryVolume;
 
     // Признак хрупкости заказа.
     Boolean fragile;
